Splits Timeline::Event classes into own files
