+++ app/models/timeline/complaint_event.rb
@@ -1,0 +1,10 @@
+class Timeline::ComplaintEvent < Timeline::Event
+  def initialize(event)
+    @name = event[0]
+    @date = Date.parse(event[1])
+  end
+
+  def label
+    Timeline::EVENT_LABELS[name] || "Updated"
+  end
+end
